projects/python_projects/abstraction-level-3/pyproject.toml
[project]
name = "abstraction-level-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "pathlib>=1.0.1",
    "pyyaml>=6.0.2",
    "typer>=0.17.4",
    "typing>=3.10.0.0",
]
